Move layout components to `src/layout`
