+++ src/components/index.js
@@ -9,15 +9,6 @@ import './stylesheets/csstrans.playlist-video.css';
 import './stylesheets/sk-loader.css';
 import './index.css';
 
-export { CTLayout } from './CTLayout';
-export { CTNavHeader } from './CTLayout/CTNavHeader';
-export { CTFragment } from './CTLayout/CTFragment';
-export { CTFooter } from './CTLayout/CTFooter';
-export { SignInMenu } from './CTLayout/CTNavHeader/NavHeaderMenu/SignInMenu';
-export { CTErrorWrapper } from './CTLayout/CTErrorWrapper';
-export { CTLoader } from './CTLayout/CTLoader';
-export { CTLoadable } from './CTLayout/CTLoadable';
-
 export { SignInPrompt } from './SignInPrompt';
 
 export { Poster } from './Poster';
