+++ src/layout/CTHeading/index.scss
