+++ src/layout/CTLayout/responsive.js
