+++ src/layout/CTNavSidebar/default-sidebar-props/admin-items.js
