+++ src/layout/CTNavSidebar/default-sidebar-props/default-items.js
