+++ src/layout/CTNavSidebar/default-sidebar-props/index.js
