+++ src/layout/CTNavSidebar/default-sidebar-props/instructor-items.js
