+++ src/screens/NotFound404/index.js
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
-import { CTErrorWrapper } from '../../components';
-import { api } from '../../utils';
+import { CTErrorWrapper } from 'components';
+import { api } from 'utils';
 
 export function NotFound404() {
   useEffect(() => {
